devcontainer support, git installation option

+++ docker-dev.sh
@@ -1,12 +1,22 @@
 #!/bin/sh
 
-DEF_TAG=$(openssl rand -hex 4)
+# Generate a random 8 digit hex number
+DEF_TAG=$(openssl rand -hex 8)
 
-sed -i 's/\r//' docker-dev-config.env
+# Colors
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+YELLOW='\033[0;33m'
+BLUE='\033[0;34m'
+NC='\033[0m'
+
+sed -i 's/\r//' docker-dev-config.env # Remove Windows line endings
 source docker-dev-config.env
 
-TAG=$(grep "^RAND=" docker-dev-config.env | cut -d'=' -f2-)
+# Get the TAG is set in the config.env
+TAG=$(grep "^TAG=" docker-dev-config.env | cut -d'=' -f2-)
 
+# if TAG is not set in the config.env, set it to a random 8 digit hex number
 if [ -z "$TAG" ]; then
     sed -i "s/^TAG=.*$/TAG=${DEF_TAG}/" docker-dev-config.env
     TAG=${DEF_TAG}
@@ -37,32 +47,54 @@ fi
 IMAGE_NAME=${PROJECT_NAME}-dev-image:${TAG}
 CONTAINER_NAME=${PROJECT_NAME}-dev-container-${TAG}
 
-cat <<EOF > Dockerfile.dev
+echo -e "Project Name: ${YELLOW}${PROJECT_NAME}${NC}\n"
+
+# Build the Docker image if not already built
+# Check if the image with the same name already exists
+if [ "$(docker images -q ${IMAGE_NAME} 2> /dev/null)" ]; then
+    echo -e "${YELLOW}Docker Image already exists. Skipping the build process...${NC}"
+else
+    read -p "Install git in the dev-container (y/n)?" GIT
+
+    cat <<EOF > Dockerfile.dev
 FROM node:20.17.0-alpine3.19
+
+RUN npm remove -g yarn || true
 RUN npm install -g corepack && corepack enable && corepack prepare yarn@4.4.1 --activate
+EOF
+
+    if [ "$GIT" = "y" ]; then
+        cat <<EOF >> Dockerfile.dev
+RUN apk add --no-cache git
+EOF
+    fi
+
+    cat <<EOF >> Dockerfile.dev
 WORKDIR ${MOUNT_PATH}
 EXPOSE ${CONTAINER_PORT}
 CMD ["/bin/ash"]
 EOF
 
-DOCKER_FILE_PATH=Dockerfile.dev
-
-echo "Project Name: ${PROJECT_NAME}"
-
-echo "Building Docker Image with Yarn 4.4.1"
-echo "Docker Image Name: ${IMAGE_NAME}"
-
-# Build the Docker image if not already built
-docker build -t ${IMAGE_NAME} -f ${DOCKER_FILE_PATH} .
-
-echo "Docker Image built successfully"
-echo "Starting Docker Container and mounting the project directory"
-
-# check if the container with the same already exists (running or stopped) remove it and start a new one
-if [ "$(docker ps -aq -f name=${CONTAINER_NAME})" ]; then
-    docker stop ${CONTAINER_NAME}
-    docker rm -f ${CONTAINER_NAME}
+    DOCKER_FILE_PATH=Dockerfile.dev
+    echo -e "${BLUE}Building Docker Image with Yarn 4.4.1"
+    echo -e "Docker Image Name:${YELLOW} ${IMAGE_NAME}${NC}"
+    docker build -t ${IMAGE_NAME} -f ${DOCKER_FILE_PATH} .
 fi
 
-# Run the Docker container
-docker run -it -e CHOKIDAR_USEPOLLING=true -e WATCHPACK_POLLING=true -p ${HOST_PORT}:${CONTAINER_PORT} --name ${CONTAINER_NAME} -v $(pwd):${MOUNT_PATH} ${IMAGE_NAME} ash
+BUILD_STATUS=$?
+
+if [ $BUILD_STATUS -eq 0 ]; then
+    echo -e "${GREEN}Docker Image built successfully${NC}\n"
+    echo -e "${BLUE}Starting Docker Container and mounting the project directory${NC}"
+    # check if the container with the same already exists (running or stopped) remove it and start a new one
+    if [ "$(docker ps -aq -f name=${CONTAINER_NAME})" ]; then
+        docker stop ${CONTAINER_NAME}
+        docker rm -f ${CONTAINER_NAME}
+    fi
+
+    # Run the Docker container
+    docker run --rm -it -e CHOKIDAR_USEPOLLING=true -e WATCHPACK_POLLING=true -p ${HOST_PORT}:${CONTAINER_PORT} --name ${CONTAINER_NAME} -v $(pwd):${MOUNT_PATH} --user $(id -u):$(id -g) ${IMAGE_NAME} ash
+else
+    echo -e "${RED}Docker Image build failed with status ${BUILD_STATUS}${NC}\n"
+    exit 1
+fi
